Move schema_ext to conftest
